Refactor/audio buffer base source node (#504)
* refactor: added AudioBufferBaseSourceNode class as a base for all sources playing AudioBuffers

* ci: yarn format

* fix: fixed range of bpm in drums

* refactor: aligned onended with web

* refactor: updated web onended event type

* chore: updated PR template

* refactor: further refactoring

* ci: yarn format

* refactor: further refactoring

* refactor: removed bufferId arg from enqueueBuffer

---------
