+++ packages/react-native-audio-api/common/cpp/audioapi/core/sources/AudioBufferQueueSourceNode.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <audioapi/core/sources/AudioBuffer.h>
-#include <audioapi/core/sources/AudioScheduledSourceNode.h>
+#include <audioapi/core/sources/AudioBufferBaseSourceNode.h>
 #include <audioapi/libs/signalsmith-stretch/signalsmith-stretch.h>
 
 #include <memory>
@@ -15,63 +15,34 @@ namespace audioapi {
 class AudioBus;
 class AudioParam;
 
-class AudioBufferQueueSourceNode : public AudioScheduledSourceNode {
+class AudioBufferQueueSourceNode : public AudioBufferBaseSourceNode {
  public:
     explicit AudioBufferQueueSourceNode(BaseAudioContext *context);
     ~AudioBufferQueueSourceNode() override;
 
-    [[nodiscard]] std::shared_ptr<AudioParam> getDetuneParam() const;
-    [[nodiscard]] std::shared_ptr<AudioParam> getPlaybackRateParam() const;
-
-    void start(double when, double offset);
     void stop(double when) override;
     void pause();
 
-    void enqueueBuffer(const std::shared_ptr<AudioBuffer> &buffer, int bufferId, bool isLastBuffer);
+    void enqueueBuffer(const std::shared_ptr<AudioBuffer> &buffer, bool isLastBuffer);
     void disable() override;
 
-    void setOnPositionChangedCallbackId(uint64_t callbackId);
-    void setOnPositionChangedInterval(int interval);
-    void sendOnPositionChangedEvent();
-
  protected:
-    std::mutex &getBufferLock();
     void processNode(const std::shared_ptr<AudioBus>& processingBus, int framesToProcess) override;
-    double getStopTime() const override;
+    double getCurrentPosition() const override;
 
  private:
-    std::mutex bufferLock_;
-
-    // pitch correction
-    std::shared_ptr<signalsmith::stretch::SignalsmithStretch<float>> stretch_;
-    std::shared_ptr<AudioBus> playbackRateBus_;
-
-    // k-rate params
-    std::shared_ptr<AudioParam> detuneParam_;
-    std::shared_ptr<AudioParam> playbackRateParam_;
-
-    // internal helper
-    double vReadIndex_;
-
     // User provided buffers
-    std::queue<std::pair<int, std::shared_ptr<AudioBuffer>>> buffers_;
-    int bufferId_ = 0;
+    std::queue<std::shared_ptr<AudioBuffer>> buffers_;
     bool isLastBuffer_ = false;
     bool isPaused_ = false;
 
-    // positionChanged event props: callbackId, update interval in frames, time since last update in frames
-    uint64_t onPositionChangedCallbackId_ = 0;
-    int onPositionChangedInterval_;
-    int onPositionChangedTime_ = 0;
-    double position_ = 0;
-
-    void processWithPitchCorrection(const std::shared_ptr<AudioBus> &processingBus,
-                                    int framesToProcess);
+    double playedBuffersDuration_ = 0;
 
     void processWithoutInterpolation(
             const std::shared_ptr<AudioBus>& processingBus,
             size_t startOffset,
-            size_t offsetLength);
+            size_t offsetLength,
+            float playbackRate) override;
 };
 
 } // namespace audioapi
