+++ packages/react-native-audio-api/common/cpp/audioapi/core/sources/AudioBufferSourceNode.cpp
@@ -13,25 +13,14 @@ namespace audioapi {
 AudioBufferSourceNode::AudioBufferSourceNode(
     BaseAudioContext *context,
     bool pitchCorrection)
-    : AudioScheduledSourceNode(context),
+    : AudioBufferBaseSourceNode(context),
       loop_(false),
+      loopSkip_(false),
       loopStart_(0),
       loopEnd_(0),
-      pitchCorrection_(pitchCorrection),
-      vReadIndex_(0.0) {
+      pitchCorrection_(pitchCorrection) {
   buffer_ = std::shared_ptr<AudioBuffer>(nullptr);
   alignedBus_ = std::shared_ptr<AudioBus>(nullptr);
-
-  detuneParam_ = std::make_shared<AudioParam>(
-      0.0, MOST_NEGATIVE_SINGLE_FLOAT, MOST_POSITIVE_SINGLE_FLOAT, context);
-  playbackRateParam_ = std::make_shared<AudioParam>(
-      1.0, MOST_NEGATIVE_SINGLE_FLOAT, MOST_POSITIVE_SINGLE_FLOAT, context);
-
-  playbackRateBus_ = std::make_shared<AudioBus>(
-      RENDER_QUANTUM_SIZE * 3, channelCount_, context_->getSampleRate());
-
-  stretch_ =
-      std::make_shared<signalsmith::stretch::SignalsmithStretch<float>>();
 
   isInitialized_ = true;
 }
@@ -57,15 +46,6 @@ double AudioBufferSourceNode::getLoopStart() const {
 
 double AudioBufferSourceNode::getLoopEnd() const {
   return loopEnd_;
-}
-
-std::shared_ptr<AudioParam> AudioBufferSourceNode::getDetuneParam() const {
-  return detuneParam_;
-}
-
-std::shared_ptr<AudioParam> AudioBufferSourceNode::getPlaybackRateParam()
-    const {
-  return playbackRateParam_;
 }
 
 std::shared_ptr<AudioBuffer> AudioBufferSourceNode::getBuffer() const {
@@ -144,34 +124,6 @@ void AudioBufferSourceNode::disable() {
   alignedBus_.reset();
 }
 
-std::mutex &AudioBufferSourceNode::getBufferLock() {
-  return bufferLock_;
-}
-
-void AudioBufferSourceNode::sendOnPositionChangedEvent() {
-  if (onPositionChangedTime_ > onPositionChangedInterval_) {
-    std::unordered_map<std::string, EventValue> body = {
-        {"value", getStopTime()}};
-
-    context_->audioEventHandlerRegistry_->invokeHandlerWithEventBody(
-        "positionChanged", onPositionChangedCallbackId_, body);
-
-    onPositionChangedTime_ = 0;
-  }
-
-  onPositionChangedTime_ += RENDER_QUANTUM_SIZE;
-}
-
-void AudioBufferSourceNode::setOnPositionChangedInterval(int interval) {
-  onPositionChangedInterval_ = static_cast<int>(
-      context_->getSampleRate() * static_cast<float>(interval) / 1000);
-}
-
-void AudioBufferSourceNode::setOnPositionChangedCallbackId(
-    uint64_t callbackId) {
-  onPositionChangedCallbackId_ = callbackId;
-}
-
 void AudioBufferSourceNode::processNode(
     const std::shared_ptr<AudioBus> &processingBus,
     int framesToProcess) {
@@ -195,7 +147,7 @@ void AudioBufferSourceNode::processNode(
   }
 }
 
-double AudioBufferSourceNode::getStopTime() const {
+double AudioBufferSourceNode::getCurrentPosition() const {
   return dsp::sampleFrameToTime(
       static_cast<int>(vReadIndex_), buffer_->getSampleRate());
 }
@@ -224,47 +176,6 @@ void AudioBufferSourceNode::processWithoutPitchCorrection(
   } else {
     processWithInterpolation(
         processingBus, startOffset, offsetLength, computedPlaybackRate);
-  }
-}
-
-void AudioBufferSourceNode::processWithPitchCorrection(
-    const std::shared_ptr<AudioBus> &processingBus,
-    int framesToProcess) {
-  size_t startOffset = 0;
-  size_t offsetLength = 0;
-
-  auto time = context_->getCurrentTime();
-  auto playbackRate = std::clamp(
-      playbackRateParam_->processKRateParam(framesToProcess, time), 0.0f, 3.0f);
-  auto detune = std::clamp(
-      detuneParam_->processKRateParam(framesToProcess, time) / 100.0f,
-      -12.0f,
-      12.0f);
-
-  playbackRateBus_->zero();
-
-  auto framesNeededToStretch =
-      static_cast<int>(playbackRate * static_cast<float>(framesToProcess));
-
-  updatePlaybackInfo(
-      playbackRateBus_, framesNeededToStretch, startOffset, offsetLength);
-
-  if (playbackRate == 0.0f || (!isPlaying() && !isStopScheduled())) {
-    processingBus->zero();
-    return;
-  }
-
-  processWithoutInterpolation(
-      playbackRateBus_, startOffset, offsetLength, playbackRate);
-
-  stretch_->process(
-      playbackRateBus_.get()[0],
-      framesNeededToStretch,
-      processingBus.get()[0],
-      framesToProcess);
-
-  if (detune != 0.0f) {
-    stretch_->setTransposeSemitones(detune);
   }
 }
